Sketch out a testable ProxyForwarder
ProxyForwarder is an http.Handler with an internal pointer to the oxy
proxying library's simple Fowarder. This is a non-loadbalanced proxy.

Every request goes through the ProxyFowarder's ServeHTTP method. The
request is inspected, and the domain is split out of the URL. We use the
domain to query the local boltdb and get the IP address of the correct
upstream. All the oxy library needs to do its work is a modified URL
field on the request.

In our tests, we only use localhost, but spin up 3 servers on random
ports. We send requests to the single proxy on port 42069, but expect
particular http status headers depending on which backend we expect to
be routed to.