Refactor state flow
On program start, we initialize a global *Store. We call functions with
that store from package api to set the data on the store we know we
need. The functions in API might make several grpc calls to load the
store. It is not a 1-1 mapping of grpc API to our frontend. Refs #3