Add client commands with a curvetls public key crypto security layer
Create a new grpcclient package for our wrapper around the generated
grpc interface.

Keys are generated by the curvetls library. A client's keypair is only
authorized if it is actually saved to the co-chair instance's database.

New config file:
* cli client commands require a file at ~/.config/co-chair/client.toml
* this config is modeled in grpcclient/client.go

New commands:
* get-client-keys - print client keypair to stdout, and add the
generated client public key to the database; must be local to co-chair
instance.
* state - print proxy config
* put - put an upstream

Makefile changes:
* add no-auth0 run config to Makefile to allow testing grpc api security
without depending on auth0. The reason for this is that auth0 protects
our web UI, not the public api.

Still TODO:
* generate keys from web UI
* key revocation