Improve the build script
build.sh has been reworked to complete a static build of frontend assets into
the co-chair binary. If parcel is detected with pgrep, frontend compilation
is avoided, otherwise invoke `parcel build` and copy the output to
frontend/static, and run assets_generate.go

We also will start checking in generated frontend code. It's not working right
now, but we've already crossed the bridge of breaking master, so yolo.